Create newtonfractal
Code to calculate roots of an equation using Newton-Raphson's algorithm.   The point is to feed it an equation which has complex roots and record, for each input point in the  x,iy plane, which root it converges to.  The plotted result can be pretty to look at.  This is an example of "Chaos Theory."